If we list all the natural numbers below 10 that are multiples of 3 or 5, we get 3, 5, 6 and 9. The sum of these multiples is 23. Write a program that computes and prints the sum of all the multiples of 3 or 5 below 1024 (excluded), followed by a new line.
    You are allowed to use the standard library

+++ functions_nested_loops/101-natural.c
@@ -26,7 +26,7 @@ int main(void)
 void sum_multiples(void)
 {
 	int i;
-	unsigned long sum = 0;
+	int sum = 0;
 
 	for (i = 0; i < 1024; i++)
 	{
@@ -36,5 +36,5 @@ void sum_multiples(void)
 			sum = sum + i;
 		}
 	}
-	printf("%lu\n", sum);
+	printf("%d\n", sum);
 }
